requirements.txt
alabaster==0.7.6
Babel==2.1.1
blinker==1.4
cffi==1.2.1
colorama==0.3.3
coverage==4.0.1
decorator==4.0.4
defusedxml==0.4.1
docutils==0.12
dominate==2.1.16
Flask==0.10.1
Flask-API==0.6.4
Flask-Babel==0.9
Flask-Bootstrap==3.3.5.7
Flask-HTTPAuth==2.7.0
Flask-Login==0.3.2
Flask-Mail==0.9.1
flask-nav==0.5
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.0
Flask-WhooshAlchemy==0.56
Flask-WTF==0.12
flipflop==1.0
guess-language==0.2
idna==2.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
mysqlclient==1.3.7
passlib==1.6.5
pbr==1.8.1
pyasn1==0.1.9
pycparser==2.14
Pygments==2.0.2
pyOpenSSL==16.2.0
python-dateutil==2.4.2
python3-openid==3.0.7
pytz==2015.6
requests==2.8.1
simplejson==3.8.1
six==1.10.0
snowballstemmer==1.2.0
speaklater==1.3
Sphinx==1.3.1
sphinx-rtd-theme==0.1.9
SQLAlchemy==1.0.8
sqlalchemy-migrate==0.10.0
sqlparse==0.1.16
Tempita==0.5.2
visitor==0.1.2
Werkzeug==0.10.4
wheel==0.24.0
Whoosh==2.7.0
WTForms==2.0.2
